# backend/api/openapi.yaml
openapi: "3.1.0"
info:
  title: SpendWise API
  description: Collaborative household financial ledger API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/workspaces/{workspaceId}/categories:
    post:
      operationId: createCategory
      tags: [Categories]
      summary: Create a category in workspace (editor+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Category name already exists in this workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: listCategories
      tags: [Categories]
      summary: List categories in workspace (viewer+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/workspaces/{workspaceId}/categories/{categoryId}:
    put:
      operationId: updateCategory
      tags: [Categories]
      summary: Update a category (editor+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/CategoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteCategory
      tags: [Categories]
      summary: Delete a category (editor+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "204":
          description: Category deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────
  # Transactions
  # ──────────────────────────────────────────
...
  /api/v1/workspaces/{workspaceId}/fundings:
    put:
      operationId: upsertFunding
      tags: [Funding]
      summary: Set funding for a user in a year-month (editor+ required)
      description: Creates or updates a funding record. Unique per (workspace, user, year_month).
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertFundingRequest"
      responses:
        "200":
          description: Funding saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FundingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    get:
      operationId: listFundings
      tags: [Funding]
      summary: List fundings in date range (viewer+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2026-01-01"
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2026-12-31"
      responses:
        "200":
          description: List of fundings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FundingResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ──────────────────────────────────────────
  # Rules
  # ──────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from `/api/v1/auth/{provider}/callback`
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "workspace not found"
      required: [error]
    Money:
      type: object
      description: Monetary amount with currency
      properties:
        amount:
          type: string
          description: Decimal amount as string for precision
          example: "150.75"
        currency:
          type: string
          default: PLN
          example: PLN
      required: [amount, currency]

    # ── Auth ──
...
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          example: Groceries
        icon:
          type: string
          example: cart
          default: ""
      required: [name]
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: string
      required: [name]
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        workspace_id:
          type: integer
          format: int64
        name:
          type: string
        icon:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, workspace_id, name, icon, created_at, updated_at]

    # ── Transactions ──
...
    UpsertFundingRequest:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
        year_month:
          type: string
          pattern: "^\\d{4}-\\d{2}$"
          example: "2026-02"
        amount:
          $ref: "#/components/schemas/Money"
      required: [user_id, year_month, amount]
    FundingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        workspace_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        year_month:
          type: string
          example: "2026-02"
        amount:
          $ref: "#/components/schemas/Money"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, workspace_id, user_id, year_month, amount, created_at, updated_at]

    # ── Rules ──
...